+++ android/capacitor/src/main/java/com/getcapacitor/Bridge.java
@@ -280,7 +280,7 @@ public class Bridge {
         webView.setWebChromeClient(new BridgeWebChromeClient(this));
         webView.setWebViewClient(this.webViewClient);
 
-        if (Build.VERSION.SDK_INT >= 24 && config.isResolveServiceWorkerRequests()) {
+        if (config.isResolveServiceWorkerRequests()) {
             ServiceWorkerController swController = ServiceWorkerController.getInstance();
             swController.setServiceWorkerClient(
                 new ServiceWorkerClient() {
@@ -293,8 +293,10 @@ public class Bridge {
         }
 
         if (!isDeployDisabled() && !isNewBinary()) {
-            SharedPreferences prefs = getContext()
-                .getSharedPreferences(com.getcapacitor.plugin.WebView.WEBVIEW_PREFS_NAME, Activity.MODE_PRIVATE);
+            SharedPreferences prefs = getContext().getSharedPreferences(
+                com.getcapacitor.plugin.WebView.WEBVIEW_PREFS_NAME,
+                Activity.MODE_PRIVATE
+            );
             String path = prefs.getString(com.getcapacitor.plugin.WebView.CAP_SERVER_PATH, null);
             if (path != null && !path.isEmpty() && new File(path).exists()) {
                 setServerBasePath(path);
@@ -346,11 +348,7 @@ public class Bridge {
 
         // Otherwise manually check WebView versions
         try {
-            String webViewPackage = "com.google.android.webview";
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                webViewPackage = "com.android.chrome";
-            }
-            PackageInfo info = InternalUtils.getPackageInfo(pm, webViewPackage);
+            PackageInfo info = InternalUtils.getPackageInfo(pm, "com.android.chrome");
             String majorVersionStr = info.versionName.split("\\.")[0];
             int majorVersion = Integer.parseInt(majorVersionStr);
             return majorVersion >= config.getMinWebViewVersion();
@@ -425,8 +423,10 @@ public class Bridge {
     private boolean isNewBinary() {
         String versionCode = "";
         String versionName = "";
-        SharedPreferences prefs = getContext()
-            .getSharedPreferences(com.getcapacitor.plugin.WebView.WEBVIEW_PREFS_NAME, Activity.MODE_PRIVATE);
+        SharedPreferences prefs = getContext().getSharedPreferences(
+            com.getcapacitor.plugin.WebView.WEBVIEW_PREFS_NAME,
+            Activity.MODE_PRIVATE
+        );
         String lastVersionCode = prefs.getString(LAST_BINARY_VERSION_CODE, null);
         String lastVersionName = prefs.getString(LAST_BINARY_VERSION_NAME, null);
 
@@ -462,9 +462,9 @@ public class Bridge {
         if (ex instanceof SocketTimeoutException) {
             Logger.error(
                 "Unable to load app. Ensure the server is running at " +
-                appUrl +
-                ", or modify the " +
-                "appUrl setting in capacitor.config.json (make sure to npx cap copy after to commit changes).",
+                    appUrl +
+                    ", or modify the " +
+                    "appUrl setting in capacitor.config.json (make sure to npx cap copy after to commit changes).",
                 ex
             );
         }
@@ -569,6 +569,9 @@ public class Bridge {
 
     public void reset() {
         savedCalls = new HashMap<>();
+        for (PluginHandle handle : this.plugins.values()) {
+            handle.getInstance().removeAllListeners();
+        }
     }
 
     /**
@@ -745,9 +748,9 @@ public class Bridge {
     private void logInvalidPluginException(Class<? extends Plugin> clazz) {
         Logger.error(
             "NativePlugin " +
-            clazz.getName() +
-            " is invalid. Ensure the @CapacitorPlugin annotation exists on the plugin class and" +
-            " the class extends Plugin"
+                clazz.getName() +
+                " is invalid. Ensure the @CapacitorPlugin annotation exists on the plugin class and" +
+                " the class extends Plugin"
         );
     }
 
@@ -822,13 +825,13 @@ public class Bridge {
             if (Logger.shouldLog()) {
                 Logger.verbose(
                     "callback: " +
-                    call.getCallbackId() +
-                    ", pluginId: " +
-                    plugin.getId() +
-                    ", methodName: " +
-                    methodName +
-                    ", methodData: " +
-                    call.getData().toString()
+                        call.getCallbackId() +
+                        ", pluginId: " +
+                        plugin.getId() +
+                        ", methodName: " +
+                        methodName +
+                        ", methodData: " +
+                        call.getData().toString()
                 );
             }
 
@@ -875,11 +878,11 @@ public class Bridge {
     }
 
     public void triggerJSEvent(final String eventName, final String target) {
-        eval("window.Capacitor.triggerEvent(\"" + eventName + "\", \"" + target + "\")", s -> {});
+        eval("window.Capacitor.triggerEvent(\"" + eventName + "\", \"" + target + "\")", (s) -> {});
     }
 
     public void triggerJSEvent(final String eventName, final String target, final String data) {
-        eval("window.Capacitor.triggerEvent(\"" + eventName + "\", \"" + target + "\", " + data + ")", s -> {});
+        eval("window.Capacitor.triggerEvent(\"" + eventName + "\", \"" + target + "\", " + data + ")", (s) -> {});
     }
 
     public void triggerWindowJSEvent(final String eventName) {