+++ cli/src/util/native-run.ts
@@ -3,6 +3,7 @@ import { dirname } from 'path';
 import c from '../colors';
 import type { PlatformTarget } from '../common';
 import { fatal } from '../errors';
+import { logger } from '../log';
 
 import { resolveNode } from './node';
 import { runCommand } from './subprocess';
@@ -44,6 +45,12 @@ export async function getPlatformTargets(platformName: string): Promise<Platform
     const err = JSON.parse(e);
     errors.push(err);
   }
+
+  if (errors.length === 0) {
+    logger.info('No devices found.');
+    return [];
+  }
+
   const plural = errors.length > 1 ? 's' : '';
   const errMsg = `${c.strong('native-run')} failed with error${plural}\n
   ${errors