+++ ios/Capacitor/Capacitor/CapacitorBridge.swift
@@ -27,6 +27,10 @@ open class CapacitorBridge: NSObject, CAPBridgeProtocol {
         #if DEBUG
         return true
         #else
+        // this is needed for SPM xcframework Capacitor.  Can eventually be removed when the SPM package moves to being source-based.
+        if let debugValue = Bundle.main.object(forInfoDictionaryKey: "CAPACITOR_DEBUG") as? String, debugValue == "true" {
+            return true
+        }
         return false
         #endif
     }
@@ -291,6 +295,7 @@ open class CapacitorBridge: NSObject, CAPBridgeProtocol {
      */
     func reset() {
         storedCalls.withLock { $0.removeAll() }
+        removeAllPluginListeners()
     }
 
     /**
@@ -564,7 +569,7 @@ open class CapacitorBridge: NSObject, CAPBridgeProtocol {
 
     func removeAllPluginListeners() {
         for plugin in plugins.values {
-			      plugin.perform(#selector(CAPPlugin.removeAllListeners(_:)), with: nil)
+            plugin.perform(#selector(CAPPlugin.removeAllListeners(_:)), with: nil)
         }
     }