+++ ios/Capacitor/Capacitor/assets/native-bridge.js
@@ -123,9 +123,8 @@ var nativeBridge = (function (exports) {
             };
         }
         else if (body instanceof FormData) {
-            const formData = await convertFormData(body);
             return {
-                data: formData,
+                data: await convertFormData(body),
                 type: 'formData',
             };
         }
@@ -657,7 +656,12 @@ var nativeBridge = (function (exports) {
                                     },
                                 });
                                 convertBody(body).then(({ data, type, headers }) => {
-                                    const otherHeaders = this._headers != null && Object.keys(this._headers).length > 0 ? this._headers : undefined;
+                                    let otherHeaders = this._headers != null && Object.keys(this._headers).length > 0 ? this._headers : undefined;
+                                    if (body instanceof FormData) {
+                                        if (!this._headers['Content-Type'] && !this._headers['content-type']) {
+                                            otherHeaders = Object.assign(Object.assign({}, otherHeaders), { 'Content-Type': `multipart/form-data; boundary=----WebKitFormBoundary${Math.random().toString(36).substring(2, 15)}` });
+                                        }
+                                    }
                                     // intercept request & pass to the bridge
                                     cap
                                         .nativePromise('CapacitorHttp', 'request', {